+++ src/sequence/seq_crdt.rs
@@ -380,16 +380,21 @@ where
     }
 
     /// Gets a list of items which are within the given indices.
+    /// Note the range of items is [start, end), i.e. the end index is not inclusive.
     pub fn in_range(&self, start: Index, end: Index) -> Option<Entries> {
-        let start_index = to_absolute_index(start, self.len() as usize)?;
-        let num_items = to_absolute_index(end, self.len() as usize)?; // end_index
+        let count = self.len() as usize;
+        let start_index = to_absolute_index(start, count)?;
+        if start_index >= count {
+            return None;
+        }
+        let end_index = to_absolute_index(end, count)?;
+        let items_to_take = end_index - start_index;
 
         self.current_lseq().map(|lseq| {
             lseq.iter()
-                .take(num_items)
-                .enumerate()
-                .take_while(|(i, _)| i >= &start_index)
-                .map(|(_, entry)| entry.clone())
+                .skip(start_index)
+                .take(items_to_take)
+                .cloned()
                 .collect::<Entries>()
         })
     }
@@ -406,7 +411,7 @@ where
 
 fn to_absolute_index(index: Index, count: usize) -> Option<usize> {
     match index {
-        Index::FromStart(index) if index as usize <= count => Some(index as usize),
+        Index::FromStart(index) if (index as usize) <= count => Some(index as usize),
         Index::FromStart(_) => None,
         Index::FromEnd(index) => count.checked_sub(index as usize),
     }